This is a complete WordPress child theme with TwentyEleven as its parent.
This repository includes all the code necessary to run this child
theme, though it does not include the parent theme; it shouldn't be too
hard to find, it comes with every WordPress install. Due to licensing
restrictions, it also does not include any images referenced in the css
and some of the php files. You will have to either roll your own, or
negotiate a license with the creators of other images on your own.